Thoughts on Common Sense by Thomas Paine
I somehow managed to go through high school and college without ever having read COMMON SENSE by Thomas Paine. I was rather delighted by it when I read it this evening.

I imagine it is difficult to read today with the same context under which it was written. You have to put yourself in a mindset of what it was like living in the American colonies in the late winter and spring of 1776, when relations with England were strained to the breaking point.

Three things impressed me while reading this book:

1. The unadorned simplicity of the writing.
2. The clarity of the argument set forth.
3. The reliance on the weight of logic and rhetoric (in a classical sense) to convey the argument for separation.

In a modern context, Paine's "pamphlet" provides a model for rhetoric largely forgotten in the modern press, but one well worth reconsidering. Issues are addressed without ad hominum attacks. Even the reply to the Quaker's call for a peaceful settlement are dealt with logically, but with grace and politeness almost unknown in today's political writing.