Thoughts on Tales of a Fourth Grade Nothing by Judy Blume
I first "read" this book in third grade. Actually, our third grade teacher read it to us, and I remember loving it. Now, thirty-something years later, I read it again, this time, to my son. We've been reading bits and pieces of it over the month of July, and I've been surprised with how much I still enjoy it.

Part of what makes the book so good is the voice of Peter Hatcher. Blume gets the voice just right, and I suppose when you are 10 years old and reading the book, Peter comes across sounding familiar, like one of your friends.

Reading the chapter where Peter's father cooks the omelette was a lot of fun. I'd forgotten that chapter entirely, and had difficulty getting through it because I was laughing so hard.

And of course, the final chapter, where we learn of Dribble's fate was still poignent. I could tell that my son felt that way too.

I'm glad I read this one again, and really glad that I got to read it with my son.