Thoughts on The Bully Pulpit by Doris Kearns Goodwin
I read this book on the heels of completing Edmund Morris's 3-volume biography of Theodore Roosevelt. Those books were a detailed look at Roosevelt's brief, remarkable life. This book took a somewhat more modest approach, looking at the lives of Roosevelt and William Howard Taft, and their interactions with the "muckraking" journalists of McLure's magazine, Ida Tarbell, Lincoln Steffens, Ray Stanard Baker, and others.

I found the parts on journalism the most remarkable parts of the book. The rise of McLure's and the writers who worked for him was, for me, a window into the evolution of journalism in the United States that I'd never before been aware of. I found myself wanting to look for the investigative pieces that Tarbell and Steffens wrote, and read them myself. It seemed to me that this was one instance in which the label "the golden age" was not hyperbole.

Goodwin has a natural flow to her writing, and reading THE BULLY PULPIT made me want to read more of Goodwin. I think that, as much as anything, is a telling recommendation for any writer. I'm looking forward to reading her book on Lincoln next.