Thoughts on Superfudge by Judy Blume
I remembered nothing at all of Superfudge when I started reading it to the Little Man. We read it on the heels of TALES OF A FOURTH GRADE NOTHING, and I have to say that this is one of those times where the sequel was not nearly as good as the original. The book attempted to live up to the reputation of the original, but fell short in the process. While there were funny parts to the book (Uncle Feather and his "bonjour stupid", for instance) it just didn't have the same feeling at the original.