Thoughts on THE NAME OF THE WIND by Patrick Rothfuss
One of the things I'd always wondered about a character like Gandalf from Tolkien's THE LORD OF THE RINGS is how he became a great wizard/warrior. At some point, Gandalf had to be a child with no notion of magic sorcery. How did he go from that child to the great wizard he became? I always thought this would make for a fascinating story.

That is, perhaps, the best way to describe what THE NAME OF THE WIND is all about. Rather than Gandalf, the story is about a boy named Kvothe who, at the outset of the story, we know to be one of the great wizard/warriors of the age, a person of legend. Now, living as a humble innkeeper "Kote", Kvothe tells his story to a Chronicler, and I finally get to see the evolution of a Gandalf-like wizard from boy to legend.

The narrative style Rothfuss uses to tell the story is a good one. The story is told first person, from Kvothe's viewpoint, as he tells his tale to Chronicler. He tells Chronicler the story will take three days. Each of these days represents one book in the series.

Much of THE NAME OF THE WIND takes place at the University where Kvothe learns magic, among other skills. The story is populated with a rich set of characters that brings the world alive. While the story is always interesting, and often exciting, it lacks the darkness and gloom of a George R. R. Martin fantasy. That said, there is a underlying hint of menace riding just below the surface throughout the story.

The story is punctuated, here and there, by returns to the present, when Kote pauses in his narrative for some reason or another. It is becoming clear that there is something equally menancing afoot in the present, and that at some point, the story of Kvothe's past and Kote's present will merge.

I enjoyed the story enough to begin the second book, THE WISE MAN'S FEAR, as soon as I finished THE NAME OF THE WIND. The third book has not yet been released, but I have a feeling if I devour the second as quickly as the first, I will be in for a frustrating wait to find out what happens next.