Thoughts on THE WISE MAN'S FEAR by Patrick Rothfuss
Too often sequels don't live up to the original, but that is not the case with Patrick Rothfuss's THE WISE MAN'S FEAR. The story continues the tale of Kvothe's growth from a young Edema Rue into a great and feared wizard. Unlike the original book, vast portions of THE WISE MAN'S FEAR take place outside the University, while Kvothe is on a kind of involuntary sabbatical. We get to see some of the world, and we get to see how Kvothe evolves through the experiences he has in that world.

In spite of everything he goes through, Kvothe does return to the University toward the end of the book, and the homecoming is understated in such a way that you, as reader, feel like you have also been away from home far too long.

There are some remarkable scenes in this book, scenes the like of which I did not believe were possible in a work of fantasy. The ambush of the group ripping off the tax collectors is one such scene; Kvothe's trial at the tree of spinning leaves is another.

I can see why people are eager for the third book to come out, but as a writer myself, I understand an empathize with the need to get the story right. There have been a few novellas in the meantime, and I am looking forward to reading each of them. In the meantime, I am eagerly awaiting the third book...