Thought on KILLING FLOOR by Lee Child
I saw an interview with Lee Child and Stephen King, after which I thought I'd give Child's first Jack Reacher book a try. I went into the book with a specific goal: to see if I could identify what makes for good action/thriller fiction.

I enjoyed KILLING FLOOR enough to buy the second book in the series. One of the things I like about the Reacher character is his isolation from mainstream society. The books reminded me of Barry Malzberg's LONE WOLF series (written as Mike Barry back in the 1970s). Although Malzberg's books are better-written in my opinion, both characters have interesting similarities.

Child is fond of fragment sentences, which I think work well in this type of writing. They lend good control of pacing. He also is fond of slowing things down at the heat of the action, so that it is almost like watching the scene unfold in slow motion.

The story was an interesting one, and if it was a little farfetched, I didn't mind because I enjoyed it anyway. It was a fun book to read because we've all wanted to get revenge against some injustice in our lives, and Reacher becomes our surrogate in this.