Thoughts on David and the Phoenix by Edward Ormondroyd
I first read DAVID AND THE PHOENIX as a boy of around 10 years old back when I lived in New England. I recall enjoying the book very much, almost unable to put it down, but I remembered very little of it when I returned to it recently. I thought it would be a good book for me and my son, who is in first grade, to read together.

It is a fun book that held my attention as grown-up, and which my son also seemed to enjoy. There are parts of it that definitely seem dated, but overall, the story was a good one, and the ending still seemed rather touching.