Thoughts on A MAN ON THE MOON by Andrew Chaikin
It had been 17 or 18 years since I first read this book in the summer of 1998. It was at least as good the second time around. The first time, I knew very little about the Apollo program beyond what most of the public knows. After that initial reading, I went on to read many other books about the Apollo program. Having come back to Chaikin's books, I remain convinced it is the best single-volume history of Apollo.