Thoughts on ON WRITING by Stephen King
This was my 5th time reading the book. I go to it when I need to refuel and remember why I write. It's just as good as always for this purpose, and the audiobook adds a particularly nice dimension. It is read by Stephen King and so, as I take my morning walks and listen to the book, it's as if King is speaking directly to me, walking with me and offering up the wisdom he's gained from his writing experience. I still think it is the best book on the craft that I have ever read.