Thoughts on HARRY POTTER AND THE SORCERER'S STONE by J. K. Rowling
We listened to the audiobook version (recently released on Audible) on the drive down to Florida this year. I only ever read the 1st book. After that, I just watched the movies, but listening to the audiobook version made me want to listen to the audiobook versions of the rest of the books as well. It was a lot of fun.