Thoughts on MY LUCKY LIFE IN AND OUT OF SHOW BUSINESS by Dick Van Dyke
Books like these are guilty pleasures for me, but I was delighted by Van Dyke's memoir. I was especially charmed by the fact that he narrated the audiobook version. It was a fun, honest look behind the curtain in Hollywood, with lots of great stories, and a few sad ones, too. I'm looking forward to the companion volume he wrote a few years later.