Thoughts on KEEP MOVING by Dick Van Dyke
KEEP MOVING was less of a memoir and more of a lessons-learned about aging from 90-year-old Dick Van Dyke. I got the sense that the audience for this book were people in their 50s and 60s, but I found it charming nonetheless. I especially enjoyed things he included like an interview he did with Carl Reiner (who at 93 is 4 years senior to Van Dyke). It was a fun book, and a fun read.