Thoughts on THE MAKING OF THE ATOMIC BOMB by Richard Rhodes
This was a much, much better book than I remembered, and my memory was that of a very good book to begin with. The ability of Rhodes to follow the begins of nuclear physics through to their conclusion, capturing the scientific and engineer drama was matched only by his ability to capture the human drama as well. I'd count this among the best nonfiction books I've ever read.