Thoughts on Abraham Lincoln: The Prairie Years and the War Years
This is the first bio of Lincoln that I finished. Years ago, toward the end of Clinton's presidency in the late 1990s, I started reading David Herbert Donald's bio of Lincoln because Clinton had mentioned it in an interview. But I never finished it. I finished this one, and it was excellent.

I never knew Lincoln to be a man of so many stories. Everything, it seemed, brought to mind a story to him. The stories were always apropos to the context and often amusing. I was fascinated by the inner workings of his office during the Civil War as well. Passing through the war, I could, as a reader, feel the weight of it press upon me, as certainly Lincoln must have.

Reading of Lincoln's assassination was difficult. Sandburg put the reader in the box, a ghost that sees everything, while Lincoln cannot. It was hard to read, knowing what was coming, and that made it powerful. The book made me want to read more about the reconstruction, what Johnson did as president, and more, what Grant did after Johnson.